Added `<% raw %>` blocks to 2 code blocks that were causing parsing problems.
This may only fix an issue with Windows, but after much investigation with installed versions of the gitbooks-cli and gitbooks itself, I could not resolve the problem without this fix. Also, it would appear that many programming books written with gitbooks use this same strategy to resolve this issue (for example, DjangoGirls), so this fix seems appropriate.

I have verified that the compiled output is what is expected.